// frontend/src/app/models/MerchantInstitutionDetail.ts
import { CountryState } from "./CountryState";
import { CountryWrapper } from "./CountryWrapper";

export interface MerchantInstitutionDetail {
  id?: number;
  address1: string;
  address2?: string;
  city: string;
  zip: string;
  country?: CountryWrapper;
  countryState?: CountryState;
  phone: string;
  fax: string;
  email: string;
}